clean up stateffects and add icon support
