+++ Content.Shared/_ArcheCrawl/StatEffects/Components/Activations/ActivateEffectOnDamageComponent.cs
@@ -1,0 +1,7 @@
+namespace Content.Shared._ArcheCrawl.StatEffects.Components.Activations;
+
+[RegisterComponent]
+public sealed class ActivateEffectOnDamageComponent : Component
+{
+
+}
