+++ Content.Shared/_ArcheCrawl/StatEffects/Components/Effects/Active/AdjustEffectStrengthEffectComponent.cs
@@ -1,4 +1,4 @@
-namespace Content.Shared._ArcheCrawl.StatEffects.Effects
+namespace Content.Shared._ArcheCrawl.StatEffects.Components.Effects.Active
 {
     [RegisterComponent]
     public sealed class AdjustEffectStrengthEffectComponent : Component
