+++ Content.Shared/_ArcheCrawl/StatEffects/Components/Effects/Active/DamageEntityEffectComponent.cs
@@ -1,6 +1,6 @@
 using Content.Shared.Damage;
 
-namespace Content.Shared._ArcheCrawl.StatEffects.Effects
+namespace Content.Shared._ArcheCrawl.StatEffects.Components.Effects.Active
 {
     [RegisterComponent]
     public sealed class DamageEntityEffectComponent : Component
