+++ Content.Shared/_ArcheCrawl/StatEffects/Components/Effects/Passive/DefenceEffectComponent.cs
@@ -1,6 +1,6 @@
 using Content.Shared.Damage;
 
-namespace Content.Shared._ArcheCrawl.StatEffects.Effects
+namespace Content.Shared._ArcheCrawl.StatEffects.Components.Effects.Passive
 {
     [RegisterComponent]
     public sealed class DefenceEffectComponent : Component
