--- Content.Shared/_ArcheCrawl/StatusEffects/Components/Activations/ActivateEffectEveryUpdateComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Shared._ArcheCrawl.StatEffects.Effects
-{
-    [RegisterComponent]
-    public sealed class ActivateEffectEveryUpdateComponent : Component
-    {
-    }
-}
