--- Content.Shared/_ArcheCrawl/StatusEffects/Components/Activations/ActivateEffectOnDamageComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Shared._ArcheCrawl.StatEffects.Effects
-{
-    [RegisterComponent]
-    public sealed class ActivateEffectOnDamageComponent : Component
-    {
-    }
-}
